// src/app/models/area.model.ts
export class Area {
    id: string;
    email: string;
    dataInsercao: string;
    car: string;
    nomeProdutor: string;
    responsavel: string;
    responsavelCpf: string;

    ap1: number;
    ap2: number;
    ap3: number;
    ap4: number;
    sistemaPlantio: number;
    totalProduzido: number;
    producaoHa: number;

    aq1: number;
    aq2: number;
    aq3: number;
    aq4: number;
    totalQueimado: number;
    porcentagemAreaQueimada: number;

    calcarioCalcitico: string;
    calcarioDolomotico: string;
    gesso: string;

    ureia: string;
    map1: string;
    map2: string;
    dap1: string;
    dap2: string;
    nitratoAmonio: string;
    uan: string;
    amoniaAnidra: string;
    sulfatoAmonio: string;
    can: string;
    ssp: string;
    tsp: string;
    kci: string;

    vinhaca: string;
    tortaFiltroBaseUmida: string;
    cinzasFuligemBaseUmida: string;

    dieselB8: string;
    dieselB10: string;
    dieselBx: string;
    dieselB20: string;
    dieselB30: string;
    biodieselB100: string;
    gasolina: string;
    etanol: string;
    biometranoTerceiros: string;
    biometranoProprio: string;
    eletricidadeRede: string;
    eletricidadePch: string;
    eletricidadeBiomassa: string;
    eletricidadeEolica: string;
    eletricidadeSolar: string;

    totalEtanolProduzidoLitros: number;
    totalMj: number;
    totalCreditoCarbono:number;
}